// src/json/schema.ts
export enum DocNodeKind {
  file = 0,
  function = 1,
  functionSignature = 2,
  enum = 4,
  enumMember = 8,
  class = 16,
  typeInterface = 32,
  typeAlias = 64,
  variable = 128,
  type = 256,
  simpleType = 257, // type | 1
  tupleType = 258, // type | 2
  objectType = 260, // type | 4
  functionTypeDocNode = 264, // type | 8
  property = 512,
  index = 1024
}

export interface DocNode {
  /**
   * May be the empty string or __unknown if not known.
   */
  name: string
  kind: DocNodeKind
}

export type DocNodeVisibility = 'public' | 'private' | 'protected'
// ...
export interface DocNodeComment {
  comment: string
  tags: DocNodeTag[]
}

/**
 * jsdoc tags, with the exceptions of @property and @param
 */
export interface DocNodeTag {
  tagName: string
  comment: string
}
// ...
export interface FunctionSignatureDocNode extends DocNode {
  kind: DocNodeKind.functionSignature
  jsdoc: DocNodeComment
  genericTypes: TypeDocNode[]
  parameters: TypeDocNode[]
  returnType: TypeDocNode
  visibility?: DocNodeVisibility
  /**
   * Only used in classes
   */
  abstract?: boolean
}

/**
 * Formatted as `[name: keyType]: type`
 */
export interface IndexSignatureDocNode extends DocNode {
  kind: DocNodeKind.index,
  jsdoc: DocNodeComment,
  keyType: TypeDocNode
  type: TypeDocNode
}

/**
 * Used within exported members to represent their type.
 */
export interface TypeDocNode extends DocNode {
  /**
   * The comment for this node, marked with @param, @property, or @prop
   */
  comment: string
  /**
   * True if the type is optional
   */
  optional: boolean
  /**
   * If applicable, the type that this type extends. Used by interfaces and potentially by generic types.
   */
  extends?: string
  /**
   * If applicable, the initializer for this node, only used by generic types.
   */
  initializer?: string
  /**
   * If applicable, generally used in object types
   */
  rest?: boolean
}